requirements.txt
aws-cdk-lib==2.196.0
constructs>=10.0.0,<11.0.0
boto3>=1.38
opensearch-py>=2.0.0
requests-aws4auth>=1.1.0